testutils/kvclientutils: move some testing utils into this new package
I'll use them in another package in a future commit.

Release note: None
